[DDW-689] Unify support request data handling
