+++ source/renderer/app/containers/loading/SyncingConnectingPage.js
@@ -3,7 +3,7 @@ import React, { Component } from 'react';
 import { observer, inject } from 'mobx-react';
 import type { InjectedProps } from '../../types/injectedPropsType';
 import SyncingConnecting from '../../components/loading/syncing-connecting/SyncingConnecting';
-import { getSupportUrl } from '../../utils/network';
+import { generateSupportRequestLink } from '../../../../common/utils/reporting';
 
 type Props = InjectedProps;
 
@@ -70,8 +70,7 @@ export default class LoadingSyncingConnectingPage extends Component<Props> {
   }
 
   handleIssueClick = async (issueButtonUrl: string) => {
-    const locale = this.props.stores.profile.currentLocale;
-    const supportUrl = await getSupportUrl(issueButtonUrl, locale);
+    const supportUrl = generateSupportRequestLink(issueButtonUrl);
     this.props.stores.app.openExternalLink(supportUrl);
   };
