+++ source/renderer/app/utils/network.js
@@ -10,28 +10,12 @@ import {
   STAGING_LATEST_VERSION_INFO_URL,
   TESTNET_LATEST_VERSION_INFO_URL,
 } from '../config/urlsConfig';
-import serialize from './serialize';
 import {
   MAINNET,
   STAGING,
   TESTNET,
   DEVELOPMENT,
 } from '../../../common/types/environment.types';
-
-const localesFillForm = {
-  'en-US': 'English',
-  'ja-JP': 'Japanese',
-};
-
-const {
-  version,
-  os,
-  apiVersion,
-  network: NETWORK,
-  build,
-  buildNumber,
-  installerVersion,
-} = global.environment;
 
 export const getNetworkExplorerUri = (network: string): string => {
   // sets default to mainnet in case env.NETWORK is undefined
@@ -72,23 +56,6 @@ export const getNetworkEkgUrl = (env: {
     ekgUrl = TESTNET_EKG_URL;
   }
   return ekgUrl;
-};
-
-export const getSupportUrl = async (baseUrl: string, locale: string) => {
-  const network = NETWORK === 'development' ? 'staging' : NETWORK;
-  const info = {
-    frontendVersion: version,
-    backendVersion: apiVersion,
-    network,
-    build,
-    installerVersion,
-    os,
-    locale,
-    product: `Daedalus wallet - ${network}`,
-    supportLanguage: localesFillForm[locale],
-    productVersion: `Daedalus ${version}+Cardano ${buildNumber}`,
-  };
-  return `${baseUrl}?${serialize(info)}`;
 };
 
 export const getLatestVersionInfoUrl = (network: string): string => {
